Refactored IOKit-specific parts of Device into a subclass
